+++ docs/release-notes.md
@@ -1,6 +1,11 @@
 # msaDocModels Release Notes
 ## Possible future features:
 
+# 0.0.87
+
+- add patterns for models TextExtractionFormatsInput
+- separate DTO model for three TextExtractionFormatsStrDTO, TextExtractionFormatsListDTO, TextExtractionFormatsDictDTO
+
 # 0.0.86
 
 - fix model queue in TextExtractionFormatsDTO